fix for crash due to comparing BYTE and int (sign difference), settings now stored in /ce/settings
git-svn-id: https://subversion.assembla.com/svn/atarijookie/trunk@470 0d2ca917-fb58-4453-bb26-a698c01b0940

+++ cosmosex/native/scsi.cpp
@@ -398,7 +398,7 @@ void Scsi::processCommand(BYTE *command)
 
     dataMedia = NULL;
 
-	BYTE attachedMediaIndex = devInfo[acsiId].attachedMediaIndex;
+	int attachedMediaIndex = devInfo[acsiId].attachedMediaIndex;
 	
     if(attachedMediaIndex != -1) {          											// if we got media attached to this ACSI ID
         dataMedia = attachedMedia[ attachedMediaIndex ].dataMedia;						// get pointer to dataMedia
